organized GetToken Proxy from handlers
